+++ lib/src/lib.rs
@@ -2,3 +2,4 @@ pub mod memory;
 pub mod options;
 pub mod cpu;
 pub mod state;
+pub mod elf;